Adding changes to grc so that it exception is not thrown when directory does not exist

+++ grc.py
@@ -52,7 +52,8 @@ for folder in data["folders"]:
     try:
         full_folder_path = os.path.join(home, folder)
         target_path = os.path.join(dotfolders_dir, folder)
-	shutil.rmtree(target_path)
+        if os.path.exists(target_path):
+            shutil.rmtree(target_path)
         print "Copying folder {} -> {}".format(full_folder_path, target_path)
         shutil.copytree(full_folder_path, target_path)
         print "\tSuccess!"
